requirements.txt
# Web Framework
Flask>=2.0.1
flask-cors>=3.0.10

# Database
SQLAlchemy>=1.4.23
Flask-SQLAlchemy>=2.5.1

# Payment Processing
stripe>=7.11.0

# Web Scraping
beautifulsoup4>=4.9.3
requests>=2.26.0
fake-useragent>=0.1.11
lxml>=5.1.0

# Environment Variables
python-dotenv>=0.19.0

# Date/Time Handling
pytz>=2024.1

# Caching
cachetools>=5.3.2

# HTTP Client
aiohttp>=3.9.3
urllib3>=2.2.0

# Security
cryptography>=42.0.2

# Utilities
typing-extensions>=4.9.0
python-dateutil>=2.8.2

# Logging
structlog>=24.1.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0

# Development Tools
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0

# Production Server
gunicorn>=21.2.0

# Performance
futures>=3.0.5
concurrent-futures>=3.0.5

# Geolocation
geopy>=2.4.1

# Image Handling
Pillow>=10.2.0
requests-html>=0.10.0
selenium>=4.18.1
webdriver_manager>=4.0.1

# Google OAuth
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
requests-oauthlib>=1.3.1

# Security
flask-talisman>=1.0.0
flask-limiter>=3.3.1
pyotp>=2.6.0
qrcode>=7.3
jwt>=1.3.1

# Monitoring
sentry-sdk>=1.4.3
elasticsearch>=7.15.0

# Caching & Queue
redis>=3.5.3
celery>=5.2.7
cachetools>=5.3.0

# Validation
marshmallow>=3.19.0

# Testing
pytest>=7.3.1
pytest-cov>=4.0.0
faker>=18.3.1

# Security Dependencies
flask-security-too>=4.1.5
flask-talisman>=1.0.0
flask-limiter>=3.3.1
argon2-cffi>=21.3.0
pyotp>=2.6.0
qrcode>=7.4.2
bleach>=6.0.0
cryptography>=41.0.1
pyOpenSSL>=23.2.0